ui: refactor platforms ui layer - add `Components` within `Platforms` directory, move all structs from the primary view file into their own components directory - Rename `Platform` to `Auth`, and rename `PlatformViews` to `Platforms` directory to more closely align with contents of the directories. - Move `PlatformViews` directory into the `Views` directory to keep all views in the `Views` directory - Convert groups to folders, because that is the new recommendation and default for xcode projects
