+++ SMSWithoutBorders-Production/Views/Platforms/EmailPlatformView.swift
