+++ SMSWithoutBorders-Production/Views/Platforms/TextPlatformView.swift
